requirements.txt
streamlit
sentence-transformers
numpy
python-dotenv
openai
langchain_community
PyMuPDF
transformers
torch
pandas
requests
arxiv
reportlab
